Move compatibility back to Java SE 1.7:
In Liferay 7.0 the process of validating LPKGs doesn‘t use the current runtime.
Instead it uses a fictitous target environment handmade in source. Refer to the
`DefaultIndexValidator` class to see where this happens.

The issue is that this target environment has Java SE 1.7 set as its `osgi.ee`
version. This differs from what is stated in the documentation for Liferay 7,
which should target Java 8.

All this prevents us to publish to the Marketplace, and therefore we are moving
back to Java SE 1.7. After all we didn’t use any Java 1.8 specific APIs.